requirements.txt
tweepy==3.10.0
requests==2.24.0
selenium==3.141.0
google==3.0.0
beautifulsoup4==4.9.3
credentials==1.1
python-dotenv==0.17.1
